#11742 send defaultValue for filters if value is not set
